+++ Source/File2Startup/Tools/Interop/Delegates.vb
@@ -93,8 +93,7 @@ Namespace DevCase.Interop.Win32
         ''' </returns>
         ''' ----------------------------------------------------------------------------------------------------
         <SuppressUnmanagedCodeSecurity>
-        Public Delegate Function EnumWindowsProc(hwnd As IntPtr,
-lParam As IntPtr) As Boolean
+        Public Delegate Function EnumWindowsProc(hwnd As IntPtr, lParam As IntPtr) As Boolean
 
         ''' ----------------------------------------------------------------------------------------------------
         ''' <summary>
@@ -125,8 +124,7 @@ lParam As IntPtr) As Boolean
         ''' </returns>
         ''' ----------------------------------------------------------------------------------------------------
         <SuppressUnmanagedCodeSecurity>
-        Public Delegate Function EnumChildWindowsProc(hwnd As IntPtr,
-lParam As IntPtr) As Boolean
+        Public Delegate Function EnumChildWindowsProc(hwnd As IntPtr, lParam As IntPtr) As Boolean
 
         ''' ----------------------------------------------------------------------------------------------------
         ''' <summary>
@@ -157,127 +155,7 @@ lParam As IntPtr) As Boolean
         ''' </returns>
         ''' ----------------------------------------------------------------------------------------------------
         <SuppressUnmanagedCodeSecurity>
-        Public Delegate Function EnumThreadWindowsProc(hwnd As IntPtr,
-lParam As IntPtr) As Boolean
-
-#End Region
-
-#Region " Hidden Base Members "
-
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <summary>
-        ''' Serves as a hash function for a particular type.
-        ''' </summary>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <returns>
-        ''' A hash code for this instance, suitable for use in hashing algorithms and data structures like a hash table.
-        ''' </returns>
-        ''' ----------------------------------------------------------------------------------------------------
-        <EditorBrowsable(EditorBrowsableState.Never)>
-        <DebuggerNonUserCode>
-        Public Shadows Function GetHashCode() As Integer
-            Return MyBase.GetHashCode()
-        End Function
-
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <summary>
-        ''' Gets the <see cref="Type"/> of the current instance.
-        ''' </summary>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <returns>
-        ''' The exact runtime type of the current instance.
-        ''' </returns>
-        ''' ----------------------------------------------------------------------------------------------------
-        <EditorBrowsable(EditorBrowsableState.Never)>
-        <DebuggerNonUserCode>
-        Public Shadows Function [GetType]() As Type
-            Return MyBase.GetType()
-        End Function
-
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <summary>
-        ''' Determines whether the specified <see cref="Object"/> is equal to this instance.
-        ''' </summary>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <param name="obj">
-        ''' Another object to compare to.
-        ''' </param>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <returns>
-        ''' <see langword="True"/> if the specified <see cref="Object"/> is equal to this instance; 
-        ''' otherwise, <see langword="False"/>.
-        ''' </returns>
-        ''' ----------------------------------------------------------------------------------------------------
-        <EditorBrowsable(EditorBrowsableState.Never)>
-        <DebuggerNonUserCode>
-        Public Shadows Function Equals(obj As Object) As Boolean
-            Return MyBase.Equals(obj)
-        End Function
-
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <summary>
-        ''' Determines whether the specified <see cref="Object"/> instances are considered equal.
-        ''' </summary>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <param name="objA">
-        ''' The first object to compare.
-        ''' </param>
-        ''' 
-        ''' <param name="objB">
-        ''' The second object to compare.
-        ''' </param>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <returns>
-        ''' <see langword="True"/> if the objects are considered equal; otherwise, <see langword="False"/>.
-        ''' <para></para>
-        ''' If both <paramref name="objA"/> and <paramref name="objB"/> are <see langword="Nothing"/>, 
-        ''' the method returns <see langword="True"/>.
-        ''' </returns>
-        ''' ----------------------------------------------------------------------------------------------------
-        <EditorBrowsable(EditorBrowsableState.Never)>
-        <DebuggerNonUserCode>
-        Public Shared Shadows Function Equals(objA As Object, objB As Object) As Boolean
-            Return Object.Equals(objA, objB)
-        End Function
-
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <summary>
-        ''' Determines whether the specified <see cref="Object"/> instances are the same instance.
-        ''' </summary>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <param name="objA">
-        ''' The first object to compare.
-        ''' </param>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <param name="objB">
-        ''' The second object to compare.
-        ''' </param>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <returns>
-        ''' <see langword="True"/> if <paramref name="objA"/> is the same instance as <paramref name="objB"/> 
-        ''' or if both are <see langword="Nothing"/>; otherwise, <see langword="False"/>.
-        ''' </returns>
-        ''' ----------------------------------------------------------------------------------------------------
-        <EditorBrowsable(EditorBrowsableState.Never)>
-        <DebuggerNonUserCode>
-        Public Shared Shadows Function ReferenceEquals(objA As Object, objB As Object) As Boolean
-            Return Object.ReferenceEquals(objA, objB)
-        End Function
-
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <summary>
-        ''' Returns a String that represents the current object.
-        ''' </summary>
-        ''' ----------------------------------------------------------------------------------------------------
-        ''' <returns>
-        ''' A string that represents the current object.
-        ''' </returns>
-        ''' ----------------------------------------------------------------------------------------------------
-        <EditorBrowsable(EditorBrowsableState.Never)>
-        <DebuggerNonUserCode>
-        Public Overrides Function ToString() As String
-            Return MyBase.ToString()
-        End Function
+        Public Delegate Function EnumThreadWindowsProc(hwnd As IntPtr, lParam As IntPtr) As Boolean
 
 #End Region